atmosphere-analyzer/backend/requirements.txt
Django==5.1.2
djangorestframework==3.15.2
django-cors-headers==4.6.0
gunicorn==23.0.0
whitenoise==6.8.2
